submissions/Brahminladka/requirements.txt
streamlit
openai-whisper
sounddevice
numpy
ciq
ninebit_ciq
